internal: move inference diagnostics to hir
